CREATE validator done. PUT and DELETE validator not done

+++ app/src/Services/AllergensService.php
@@ -4,6 +4,7 @@ namespace App\Services;
 
 use App\Core\Result;
 use App\Models\AllergensModel;
+use App\Validation\Validator;
 
 
 class AllergensService
@@ -14,26 +15,83 @@ class AllergensService
     {
         //TODO: 1) Validate the received resource data about the new resource to be created.
         //* VALIDATE USING VALITRON
+
+        $rules = array(
+            'allergen_id' => [
+                ['regex', '/^[A-Z][0-9]{2}$/']
+            ],
+            'allergen_name' => [
+                'required',
+                'ascii',
+                array('lengthMin', 4)
+            ],
+            "allergen_reaction_type" => [
+                'required',
+                array('lengthMin', 4)
+            ],
+            "food_group" => [
+                'required',
+                array('in', ["Fruits", "Vegetables", "Pulses", "Grains", "Proteins", "Dairy", "Fats and Oils", "Sweets and Snacks", "Beverages"])
+            ],
+            "food_type" => [
+                'required',
+                array('lengthMin', 4)
+            ],
+            "food_origin" => [
+                'ascii',
+                array('lengthMin', 4)
+
+            ],
+            "food_item" => [
+                'required',
+                'ascii',
+                array('lengthMin', 4)
+            ]
+
+        );
         //! RETRURN RIGHT AWAY AS SOON AS YOU DETECT ANY INVALID INPUTS
-        if (1 == 0) {
-            return Result::failure("Error!", [
-                "allergen_name"  => "The following allergen name is invalid",
-                "allergen_type" => "The allergen type is invalid",
-            ]);
-        }
+
         //TODO: 2) Insert the resource into the DB table
         //* We can use an array and using the first 1 so we can make our lives easier
         $new_allergen = $new_allergens_info[0];
-        //  dd($new_allergen);
+        $validator = new Validator($new_allergen, [], 'en');
+        $validator->mapFieldsRules($rules);
+
+
+        if (!$validator->validate()) {
+
+            echo $validator->errorsToString();
+            // echo '<br>';
+            echo $validator->errorsToJson();
+            return Result::failure("error!");
+        }
         $last_inserted_id =  $this->allergens_model->insertAllergen($new_allergen); //
-        // Result pattern is implemented
-        // $last_insert_id = 29;
-        // Return a successful result
         return Result::success("The allergen has been created successfully!", $last_inserted_id);
     }
 
     function deleteAllergens(array $condition): Result
     {
+        if (!isset($condition['allergen_id'])) {
+            return Result::failure("Allergen ID is required");
+        }
+        //TODO: validate the id
+        echo "Service";
+        //dd($condition);
+        // $validator = new Validator(['regex', '/^[A-Z][0-9]{2}$/']);
+        // $validator->rule('allergen_id', 'regex', '/^[A-Z][0-9]{2}$/');
+        $validator = new Validator([
+            'allergen_id' => ['regex' => '/^[A-Z][0-9]{2}$/']
+        ]);
+
+        echo (!$validator->validate());
+
+        if (!$validator->validate()) {
+            //     echo $validator->errorsToString();
+            //    // echo '<br>';
+            echo $validator->errorsToJson();
+            return Result::failure("Allergen ID is not valid");
+        }
+
         $rowsDeleted = $this->allergens_model->deleteAllergen($condition);
 
         if ($rowsDeleted <= 0) {
@@ -44,7 +102,49 @@ class AllergensService
 
     function updateAllergen(array $data, array $condition): Result
     {
-        // echo $data;
+        $rules = array(
+            'allergen_id' => [
+                ['regex', '/^[A-Z][0-9]{2}$/']
+            ],
+            'allergen_name' => [
+                'required',
+                'ascii',
+                array('lengthMin', 4)
+            ],
+            "allergen_reaction_type" => [
+                'required',
+                array('lengthMin', 4)
+            ],
+            "food_group" => [
+                'required',
+                array('in', ["Fruits", "Vegetables", "Pulses", "Grains", "Proteins", "Dairy", "Fats and Oils", "Sweets and Snacks", "Beverages"])
+            ],
+            "food_type" => [
+                'required',
+                array('lengthMin', 4)
+            ],
+            "food_origin" => [
+                'ascii',
+                array('lengthMin', 4)
+
+            ],
+            "food_item" => [
+                'required',
+                'ascii',
+                array('lengthMin', 4)
+            ]
+        );
+
+        $validator = new Validator($data, [], 'en');
+        $validator->mapFieldRules($data, $rules);
+
+        if (!$validator->validate()) {
+            //     echo $validator->errorsToString();
+            //    // echo '<br>';
+            //     echo $validator->errorsToJson();
+            return Result::failure("Data is not valid");
+        }
+
         $rowsUpdate = $this->allergens_model->updateAllergen($data, $condition);
         if ($rowsUpdate <= 0) {
             return Result::failure("No row has been updated");
